PD-1350 Update App Versioning and Tutorial _index File
This commit updates the AppsUnversioned.md snippet content by removing references to version tutorials, and moving the user update contributions out of the snippet to the _index.md file for app tutorials and the UI ref content.
It makes minor updates to the tutorials Apps/_index.md file.
After Docker Compose changes are available in the UI, the content in these index files will change in a different PR.