PD-1350 Add AppsUnversiond.md to UI Ref _index
This commit adds the AppsUnversioned.md snippet and the community contribution blerb removed from the snippet to the UIReference/Apps/_Index.md file.
The community contribution blerb should be snippetized in context of the Community Contribution content. Need a separate ticket to evaluate this content and the best approach for it.