PD-1348 Update Apps UI Ref Article
This commit updates the UI Ref /apps/_index.md article with an alias for the CustomAppsScreens.md article and deletes that article.
It removes the UI ref content for apps and points users to 24.04 content for now as 24.10 could be radically different. Updates to this article are expected after the EE UI is availble.