PD-1355 Update Contributing Article Content
This PR updates all the articles in the /Contributing/ section of the docs website.
It adds new content to the _index.md files for Issue Reporting and Documentation.
It adds new tutorial content for contributing to apps as part of the open source TrueNAS Charts, Community, and Apps projects, as well as contributing new documentation.
It reorganizes the navigation of these sections for better logical flow by subject.
It adds a new app tutorial template.

Note. After the Docker Compose UI changes make it into the UI after 8/11/2024, a new PR will update the content in the articles and templates to reflect the changes.