PD-1290 Update Admin User to Truenas_Admin
This PR updates article and snippet content containing instructions referring to the admin user account changing it to truenas_admin.
Update the installer screenshots to reflect changes in screens in articles and snippets.
This change applies to 24.10 only. Updated mentions of clean install and UI upgrade scenarios where the admin user name is retained.