PD-1356 Update Apps UI Ref Content
This commit updates the UI ref content for the application common screens, removes content from the Apps tutorials _index.md file and creates a new snippet, then adds that snippet to both the UI ref and tutorial index files. It updates some of the UI images.  More UI changes are expected in nightly BETA.1 builds so screenshot commits are planned for this PR.