PD-1356 Update with BETA.1 UI Changes
This commit changes content and screenshots based on the changes added to the BETA.1-INTERNAL-1 build.
Charts train is now just Stable.
Settings menu has Manage Container Images restored as an option.
Edit Catalog screen shows checkboxes for the trains now. It is not showing the enterprise train yet, shows as test for now. Expected changes in RC1 to add enterprise train.