PD-1300 Changes to the MinIO Tutorials
This commit tacks on the changes to the MinIO tutorial, snippets and screenshots, requested in PD-1300 to the PD-1352 PR.  It updates the configuration instructions and UI sections with current information. It updates the apps/charts/minio/_index.md file and all the snippets in this article correcting incorrect or misleading storage configuration content, and removing what no longer applies.