Remove CLI Ref guide from 24.04 docs (obsoleted)
This content is considered not recommended for general use and the feature is not currently planned to receive any updated during the 24.04 lifecycle. As such, I'm removing the content, aliasing deleted articles back to a single URL that notes the the content was obsolete and thus removed, and build tested to remove any references to the deleted articles.