PD-1363 Update Network Content
This PR moves content from ManagingInterfaces.md to the /network/interfaces/_index.md file and deletes the ManagingInterfaces.md article.
It adds an alias for the deleted article to the _index.md file.
It updates the content with the testing network changes content.
It updates the UI ref /network/interfaceScreens.md article content and applies new formatting for settings with multiple options.
It updates broken links across numerous articles that linked to the deleted article.