PD-1375 Update Syncthing Enterprise Tutorial
This commit updates the content in the /Apps/EnterpriseTrain/Syncthing.md tutorial.
Adds new screenshots of the updated screen.
Updated configuration instructions in the Syncthing.md artricle and related snippets. The storage configuration section is updated to include the two datasets, home and data1, this app now requires. Port the storage instructions to the stable app version of syncthing.