PD-1402 Move /Apps top level articles
This commit creates the new /TruenasApps folder/top navigation for all SCALETutorials/Apps articles.
This commit fixes links, adds aliases, replaces SCALE with TrueNAS mentions in the article/

It updates the Sandboxes.md article content with a Before you Begin section with information on logging in as root, which is required for the commands in the article sections. It makes other minor changes to the article related to style guidelines.