PD-1375 Updates to Main Apps Index.md
this commit makes changes to the the /truenas_apps/_index.md file.

It corrects information and mentions of the ix-apps dataset, when it gets created, and when the old ix-applications dataset exists in 24.10.

It adds comments about the change from kubernetes backend to docker compose and links to the 24.10 release notes.