PD-1403 Remove CORE articles, snippets, and Images
This commit deletes the CORE articles, snippets, and images.
It updates the advancedscheduler.md snippet with a SCALE screenshot to replace the path to the CORE image and the image

Images remaing in the /static/images/core folder are used in SCALE iSCSI articles.

(cherry picked from commit 1fd0115f7d6985fe4085306827795243bf9f1434)