PD-1401 Update MinIO Apps
This commit updates the stable train version of the MinIO app tutorial, related snippets, the Enterprise MinIO app tutorial with changes to the resource configuration snippet, and the Syncthing Enterprise app tutorial with the changes to the resource configuration snippet.
The new AppsInstallWizardResourcesConfig.md snippet can be used in all app tutorials now and replaces the SyncthingWizardResourceConfig.md snippet.