Add AI-Powered Search widget to Docs Hub
This is an embedded widget that is a front end to DocsBot and GPT-4 based LLM query responses.

The bot is initially ingested with public documentation about TrueNAS from the Docs Hub, TrueNAS.com, and a few other Open Source licensed software projects that are relevant to TrueNAS.

Responses are experimental and should be used cautiously.

Community feedback and tuning suggestions are welcome!