PD-1413 Update MinIO and Syncthing Tutorials
This commit change snippets used in these tutorials, breaks up the AppInstallWizardNetworkConfig.md snippet, moving certificate information out and into a separate snippet.
It updates the AppWizardStorageSMBOption.md snippet, removing the migrate mode info out of the snippet and into the Syncthing enterprise tutorial as it is the only app with this setting option, and it expands the content in that snippet.