PD-1401 / 25.04 / Pd 1401 update min io stable and enterprise articles (#3048)
* PD-1401 Update MinIO Apps

This commit updates the stable train version of the MinIO app tutorial, related snippets, the Enterprise MinIO app tutorial with changes to the resource configuration snippet, and the Syncthing Enterprise app tutorial with the changes to the resource configuration snippet.
The new AppsInstallWizardResourcesConfig.md snippet can be used in all app tutorials now and replaces the SyncthingWizardResourceConfig.md snippet.

* Revert "PD-1401 Update MinIO Apps"

This reverts commit fd6a8ea4734cf2141c8f18cb9522a3dfe234be54.

* PD-1401 Revert Commit and Fix

this commit reverts the commit that deleted the MinIO cluster article and fixes the content in other articles to reflect the desired changes.

* PD-1401 Update MinIO Stable Articles

This commit updates the stable train MinIO article content, creates new snippets to use across all stable and community apps, updates the MinIO and Syncthing enterprise app tutorias with a new Resource Config section snippet that can be used in all other stable or community apps.

* Update MiniIOS3AppInfoScreen.png

* PD-1401 MinIO Enterprise Tutorial Updates

This commit adds the ACL content to the enterprise tutorials, updates storage content, copies the storage configuration information into the MNMD and SNMD tutorials.
Need new screenshots after Stavros' ACL PR/commit hits the nightly build.

* PD-1401 Update two snippets

This commit updates the AppInstallWizardResourceConfig.md snippet by removing reference to Syncthing and changing how the values are entered without the suffix.
And updates the AppInstallWizardACLConfiguration.md snippet with feedback changes.

* PD-1401 Update Snippets Fix Build Error

Minor updates to snippet files, and fix the syntax issue breaking the local build

* PD-1401 Fix issue breaking build

* PD-1401 Fixed Build issue

* PD-1401 Optimize jpg

* PD-1401 Update and Compress Images

* PD-1401 Update MinIO Enterprise Articles with ACL/ACE

This commit adds new screenshots and modifies the storage config instructions to focus on the host path option with Enable ACL selected.
It updates the three tutorials with the same content and updates images used and content in the ACL Understanding wizard settings section.

* PD-1401 Add ProposeContentChange snippet to MinIO index

This commit adds the ProposeContentChange.md snippet to the stable MinIO app index.md file

* PD-1401 Update MinIO Apps

This commit updates the stable train version of the MinIO app tutorial, related snippets, the Enterprise MinIO app tutorial with changes to the resource configuration snippet, and the Syncthing Enterprise app tutorial with the changes to the resource configuration snippet.
The new AppsInstallWizardResourcesConfig.md snippet can be used in all app tutorials now and replaces the SyncthingWizardResourceConfig.md snippet.

* Revert "PD-1401 Update MinIO Apps"

This reverts commit fd6a8ea4734cf2141c8f18cb9522a3dfe234be54.

* PD-1401 Revert Commit and Fix

this commit reverts the commit that deleted the MinIO cluster article and fixes the content in other articles to reflect the desired changes.

* PD-1401 Update MinIO Stable Articles

This commit updates the stable train MinIO article content, creates new snippets to use across all stable and community apps, updates the MinIO and Syncthing enterprise app tutorias with a new Resource Config section snippet that can be used in all other stable or community apps.

* Update MiniIOS3AppInfoScreen.png

* PD-1401 MinIO Enterprise Tutorial Updates

This commit adds the ACL content to the enterprise tutorials, updates storage content, copies the storage configuration information into the MNMD and SNMD tutorials.
Need new screenshots after Stavros' ACL PR/commit hits the nightly build.

* PD-1401 Update two snippets

This commit updates the AppInstallWizardResourceConfig.md snippet by removing reference to Syncthing and changing how the values are entered without the suffix.
And updates the AppInstallWizardACLConfiguration.md snippet with feedback changes.

* PD-1401 Update Snippets Fix Build Error

Minor updates to snippet files, and fix the syntax issue breaking the local build

* PD-1401 Fix issue breaking build

* PD-1401 Fixed Build issue

* PD-1401 Optimize jpg

* PD-1401 Update and Compress Images

* PD-1401 Update MinIO Enterprise Articles with ACL/ACE

This commit adds new screenshots and modifies the storage config instructions to focus on the host path option with Enable ACL selected.
It updates the three tutorials with the same content and updates images used and content in the ACL Understanding wizard settings section.

* PD-1401 Add ProposeContentChange snippet to MinIO index

This commit adds the ProposeContentChange.md snippet to the stable MinIO app index.md file

* PD-1401 Move App Snippets to Apps subfolder

This commit changes the path for app-related snippets to include /apps in the path, and moves these snippets to the /apps subfolder

* PD-1401 Change App Snippet Paths

This commit changes more snippet paths in app articles.

* PD-1401 Update MinIO content with missing content

This commit updates the /minioApps/_index.md and MinioClustering.md articles with changes made in the PD-1413 branch. It adds the three new snippets and adds the shortcodes for them to these tutorials. It makes other changes identified such as removing duplicate content.

* PD-1401 Add missing images

This commit adds two snippets, updates image links, and adds missing images in the MinIO articles

* PD-1401 General Corrections

This commit fixes spelling issues and makes minor corrections

* Update MinIOCompleteInstall.md

correct spelling

* PD-1401 Fix Image of Enterprise ACL configuration

This commit adds a new image for the MinIO Enterprise Storage ACL and ACE configuration to the main index and each of the cluster articles MNMD and SNMD articles.

* review changes

---------

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>
Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>