PD-1413 Clean Up Snippets and Tutorials
This commit removes duplicated content, places where the snippet is called twice, and adds the correct snippets in places where the snippet called was not the right choice.
It adds a few new snippets for both the stable and enterprise versions of MinIO.
Adds a new snippet on preparing to install that we can use globally in all app tutorials.