PD-1413 / 25.04 / Pd 1413 reorganize update stable app tutorials (#3076)
* PD-1413 Update Apps UI Ref Content

This commit updates the content in the UI ref /apps/_index.md article with new screenshots and content reflecting changes made to the UI for RC1

* PD-1413 Update AppInstallWizardSettings.md and App index.md files

This commit updates the content of the AppInstallWizardSettings.md snippet, and each of the Truenas_apps _index.md file to include this snippet.
the main _index.md file replaces the expand sections with this snippet.

* PD-1413 Add ProposeArticleChange.md Snippet to Index files

This commit adds the ProposedArticleChange.md snippet to the Truenas_apps main index pages.

* PD-1413 Move App Snippets Update Paths

This commit makes the same change moving app snippets to a subfolder /apps and updating article paths for snippets.
It also updates the Syncthing.md article deleting the old snippet and replacing it with the more generally names snippet for resource config.

* PD-1413 Update MinIO Stable with new Snippets

This commit updates the stable MinIO app tutorial with new snippets that are more universal so they can be used in other app tutorials. It adds missing images found in the PD-1401 pr.

* PD-1413 Update MinIO and Snippet Content

This commit syncs changes in PD-1401 with the same files in the PD-1413 branch.

* PD-1413 Update Syncthing Create New Snippets

This commit creates new global snippets to use in app tutorials, and updates the Syncthing.md tutorial for the enterprise app. The new snippets can be used in the stable syncthing app tutorial.

* PD-1413 Update MinIO and Syncthing Tutorials

This commit change snippets used in these tutorials, breaks up the AppInstallWizardNetworkConfig.md snippet, moving certificate information out and into a separate snippet.
It updates the AppWizardStorageSMBOption.md snippet, removing the migrate mode info out of the snippet and into the Syncthing enterprise tutorial as it is the only app with this setting option, and it expands the content in that snippet.

* PD-1413 Synthing Edits More Snippets

This commit crates the advanced DNS options snippet,

It also makes other changes to MinIO and Syncthing tutorials to use the new and changed snippets.

* Update words-to-ignore.txt

* PD-1413 fix typo in Syncthingstable.md

* PD-1413 Compress Image

* PD-1413 Fix Image in MinIO Enterprise Index

This commit adds the correct image for the data1ACLandACE setting to the main MinIO Enterprise index file.

* PD-1413 Clean Up Snippets and Tutorials

This commit removes duplicated content, places where the snippet is called twice, and adds the correct snippets in places where the snippet called was not the right choice.
It adds a few new snippets for both the stable and enterprise versions of MinIO.
Adds a new snippet on preparing to install that we can use globally in all app tutorials.

* Apps images compression pass

* Update UsingTrains.md

* Apply suggestions from code review

These suggestions fix issues that didn't get pushed to the PR

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>

* PD-1413 PR Reveiw Fixes

This commit fixes the issues with the image for train settings and updated the UsingTrains.md article and the mentions in the UiRef/Apps/_index.mde articles.
Removed calls to the MinIOEnterpriseMultiModeConfig.md snippet and deleted this unnecessary snippet.

* Update words-to-ignore.txt

* Re-Add missing image

* Fix another image call and one unordered list change

---------

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>