PD-1446 PR Feedback Updates
This commit removes the section on Verifying Name Resolution based on feedback received from Andrew who suggested Ping is not the best tool and that we should provide specific dig commands to look up AD SRV records. We don't want users entering Linux commands so this section is removed.
Also corrected the information on the Domain Account Name setting.