H-Series test image for DocsBot ingest test
This is a basic export of one of the images present in the H-Series User Manual with additional metadata and keywords added.
The test is to see if adding this .png to the repository allows the Docs Bot to recall the image for responses, even if the image isn't used in any published article on the website.