PD-1463 Update Apps Content
This commit fixes issues with snippets not rendering due to pathname issues.

Adds the new image for deleting an app that includes the option to delete the ixVolume to the main apps index and updates the instructions.

Adds information about using the Force option. Not selecting this prevents apps from updating, while selecting it allows writing to the existing storage volume and permits the update to occur.

Fixes issues in the enterprise tutorial where a screenshot is duplicated.

Allowing conversion of the ixVolume to a host path/dataset is not a function yet so additional content changes might be required.

The Custom App content is under development in a separate PR.
Adds