PD-1469 Update Network Interface Content
This commit combines the content in the SettingUpStaticIPs.md with the content in the /network/interfaces/_index.md article and deletes the separate static IP article.
It deletes the AliasOrStaticIP.md snippet as it is not needed now.
It updates the procedure content to reflect the current test changes process and deleted the old screenshots.
It reorganizes content for improved readability.