PD-1518 / 25.04 / PD-1518 Update Getting Started Storage and Share Articles (#3150)
* PD-1518 Update Getting Started Storage and Share Articles

This PR updates the content in the SetUpStorageSCALE.md article with clarifications and a section on adding the dataset and share at the same time.
It updates the SetUpSharing.md article by clarifying how to either select an existing  dataset or creating a new dataset while adding a share.

* Update SetUpStorageScale.md

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>
(cherry picked from commit 15a447c2d609422992706290e5e5acdb36939e34)