PD-1499 / 25.04 / Pd 1499 create do not create root dataset share warning (#3200)
* PD-1499 Create New Snippet Warning

This PR creates a new snippet for a warning to not create a share for the root or pool-level dataset, and adds it to all SMB and NFS share articles.

* PD-1499 Add Snippet to Getting Started Sharing Article

This commit adds the new snippet to the /gettingstarted/configure/SetUpSharing.md articl

And updates the type of hint used in the snippet.

* Update static/includes/RootLevelDatasetShareWarning.md

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>

* Update static/includes/RootLevelDatasetShareWarning.md

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>

* Update static/includes/RootLevelDatasetShareWarning.md

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>
(cherry picked from commit fe6c275c35c7a47972f972ea6722bc1a96424ee4)