PD-1626 Add Redis Password Characters and User Clarification Information
This commit adds the characters not allowed in Redis passwords, and clarifies information about the TrueNAS users added as an administrator for the app, and adds content linking this article to with the Nextcloud plugin article in /solutions/integrations/nextcloud.md article.
It updates the content in two snippet files.