PD-1626 Update Plex and Nextcloud Tutorials
This commit creates a new snippet BeforeYouBeginAddAppDatasets.md for the create dataset bullet point, and removes the add user and add dataset bullet points from the BeforeYouBeginStableApps.,md snippet.
It adds html code for indented paragraphs with the <dif></dif> tags to get the expand to indent, and replaces the markdown emphasis with HTML tags.
It updates the Plex tutorial to use the new snippet.