(After Redirects) remove nightly build core docs
These are web server redirecting to the 13.0 or 13.3 docs now, as further docs updates are anticipated in those locations. With the redirects in place, we can remove the now obsoleted content locations.

Do additional cleanup on master branch unversioned references to CORE nightly docs and replace with SCALE nightly or 13.0 links as needed.