PD-1556 / 25.04 / Pd 1556 revise app tutorial template (#3292)
* PD-1556 Update Apps Contribution and Tutorial Template

This PR updates the apps tutorial template and changes embedded instructions to commented-out content instead of using code blocks that didn't render well.
It also updates the contributing article to provide more instructions on using the tutorial template.

* PD-1556 Fix formatting issues

* Update _index.md

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

fix copy/paste error in line 160

* Update _index.md

Modified the description for the BeforeYouBeginStableApps.md snippet, removing the new user and add dataset bullet points, and added the new snippet BeforeYouBeginAddAppDatasets.md to the list. This snippet is created in another PR, PD-1626 that modifies the Nextcloud tutorial.

* PD-1556 Update Contributin and Template

This commit updates the Contributing section _index.md content with the new snippets.
It updates the template with new/improved content

* PD-1556 Updates to the Contributin _index.md file

* Update AppArticleTemplate.md

Add missing /expand tag to line 263

* Update AppArticleTemplate.md

A few changes made to the template. Added internal link to storage configuration settings 
and added new commented-out text related to the InstallWizardStorageConfig2.md snippet.

* PD-1556 Add new InstallWizardLabelsConfiguration.md Snippet

* PS-1556 Add new Labels Configuration Information to the index and tempalte files

This commit updates the _index.md file table of snippets to include the new labels configuration snippet.
Updates the template with the section for Labels Configuration and the content for that section.

* Update AppArticleTemplate.md

fix missing quotation mark

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

* Update AppArticleTemplate.md

* PD-1556 Adjustments and Style Fixes

* PD-1556 Update Template

This commit updates the app template.md and InstallAppStorageConfig.md snippet files.
The snippet now includes information on the new option to create a dataset while in the app installation wizard.

* Update AppArticleTemplate.md

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>