PD-1803 / 25.10 / Clarify Apps Support Tiers (#3543)
* PD-1803: Clarify Apps Support Tiers

Adapt guidance to a standalone Support Definitions article and add Enterprise-specific notes to the landing page for each of the Apps trains that clarifies what tier of support is available to Enterprise customers for that particular type of App.

* Update words-to-ignore.txt

* Improve Apps-Community-Support.md

Adds https://fontawesome.com/v4/icon/life-ring as an svg sprite and reworks the hint shortcode using it

* Update content/TruenasApps/Support-Definitions.md

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>

* PD-1803: review comments

Slightly update community support scope with link to Github TrueNAS apps repo issues tracker.

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>