PD-1763 / 25.10 / Pd 1763 update system general articles (#3571)
* PD-1763 Update General Settings Articles

This commit pushes change and updates to the GeneralSettingsScreens.md article screenshots and setting descriptions.
It moves the feedback window tutorial content out of the snippet and deletes FeedbackWindow.md that is not used in any other article.
It moves the managing general settings content out of the GeneralSettings.md article and adds it to the mostly empty /GeneralSettings/_index.md file, an deletes the GeneralSettings.md file.
Updates the SettingUpSystemEmail.md.
Deletes unused snippets EmailOptions.md, EmailOptionsSMTP.md, EmailOptionsGmail.md, and EmailOptionsOutlook.md

* PD-1763 Minor Updates, Fix broken Links

* PD-1763: compress images

* Update SettingUpSystemEmail.md

* Update GeneralSettingsSCALESupportLicenseComplete.png

* Update GeneralSettingsSCALESupportLicenseComplete.png

* Update GeneralSettingsScreens.md

---------

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>