PD-1789 / 25.10 / Pd 1789 update smb article content (#3550)
* PD-1789 Update SMBSharesScreens.md

This commit updates setting and screen element descriptions and functions in the SMBSharesScreens.md and SMBPuroposePresets.md snippet file.

* PD-1789 Update Datasets/_index.md

This commit updates the Datasets Screens UI ref article with new content related to SMB and Multi-protocol share changes and makes grammatical changes.

* PD-1789 Update the Multi-mode Share Tutorial

This commit updates the content in the MixedModeShares.md tutorial with the new instructions for adding the share and dataset from the Add Dataset and Add SMB screens. It also makes grammatical and spelling corrections.

* PD-1789 Update the main SMBShare/_index.md tutorial

This commit updates the content in the main /shares/SMB/_index.md tutorial with the new instructions for creating the dataset with the share on the Add SMB screen and makes grammatical changes to the content.

* PD-1789 Rename and Rewrite Home Share tutorial

This commit renames the AddSMBHomeShare.md to SMBPrivateDatasetShare.md, and rewrites the content to focus on the private dataset and share configuration process to address issues raised the PD-1691 ticket.

* PD-1789 Update images and fix broken links

This commit fixes broken links to the renamed article to point to this article, and replaces an image

* Update SMBPrivateDatasetShare.md

Correct line 20 to indicate planned support is ending.

* PD-1789 Correct ACL Entry Information

This commit updates the ACL permissions instructions with when to add an ACL entry for a user or group.

* PD-1789 Remove Snippet that doesn't exist

This commit removes the the 24.04HomeDirectory.md snippet that no longer exists

* PD-1789 Remove reference to Virtualization Article

Virtualization articles are removed pending new tutorial for Instances, and VMs through instances.

* Update MixedModeShares.md

Rewords line 104 for better clarity

* Update MixedModeShares.md

* Update SMBPrivateDatasetShare.md

* Update _index.md

* Update _index.md

* Update SMBSharesScreens.md

* Update SMBPurposePresets.md

* PD-1783 Fix broken link

---------

Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>