PD-1846: Add dynamic App Versioning shortcode
This is intended to help fill an information gap for Apps available from https://github.com/truenas/apps/.

This is a shortcode that is intended for dynamic file load from a Github repository (Thanks Greg Hilston for your article from https://www.greghilston.com/post/hugo-display-dynamic-github-content/ !)

I've modified the shortcode a bit to place the raw file contents behind an html expandable element and to dynamically load some specific values from the json file as the primary output of the shortcode. This is to better tailor the output for use with TrueNAS Apps docs.

I expect the PR to be in a draft state until any additional internal requests/requirements for this change are resolved.