PD-1802 / 25.10 / Pd 1802 add iscsi info to migrate prep.md in master (#3624)
* PD-1802 Add iSCSI Migration Info and Warning

This PR adds emphatic instructions for Enterprise customers with iSCSI and/or fibre channel deployments.
It also adds some information for community users about the two major differences in SCALE iSCSI.
A second PR to add this information to 25.04 will cover what applies in that release (no fibre channel references).

* PD-1802 Minor Change

* PD-1789 Minor structure change

This commit moves the Enterprise content in the checklist under the community information.

* Update MigratePrep.md

---------

Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>