PD-1753 / 25.10 / Pd 1753 add preparing install media option instructions (#3663)
* PD-1753 Create Ventoy Snippet File

This commit creates a Ventoy snippet to include in the InstallingSCALE.md article.

* PD-1735 Update InstallingSCALE.md

This PR updates the InstallingSCALE.md article with additional options to create bootable media, and adds two new snippets: Rufus.md and Ventoy.md with details on these options.

* PD-1753 Fix style/formatting issue

Fixes the close tag missing the / in the <code> tag

* Update InstallingSCALE.md

* Update Rufus.md

---------

Co-authored-by: tonyriv3 <75626853+tonyriv3@users.noreply.github.com>