PD-1683 Update Secret Seed Content and Add new Snippet
This commit adds a new snippet file SecretSeed.md that explains what this is and why to use it, and also how to reset the login password you have not included it, with instructions to check all system settings and test to make sure everything works after uploading the config file without it.
It updates content in several articles that include the secret seed information.
It fixes spelling issues and updates the GeneralSettingsScreens.md UI ref article. It changes non-UI ref content from tutorial style to UI ref style.