PD-1266 / 13.0 / PD-1266 Update Rsync Task Content (#3672)
* PD-1266 Update Rsync Task Content

This PR updates the content in the current rsync task article in general, and updates the SSH mode instructions to what is tested and verified to work.

* PD-1266 Corrections and Updates to the Instructions

This commit includes new command examples and corrections to the procedure after testing it.

* PD-1266 Update Rsync Module Mode Instructions

This commit makes changes to the module mode rsync task instructions per PD-1742

* PD-1266 Fix Formatting issue

* Update content/CORETutorials/Tasks/CreatingRsyncTasks.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>