PD-1683 / 25.10 / Pd 1683 update config file secret seed content (#3668)
* PD-1683 Update Secret Seed Content and Add new Snippet

This commit adds a new snippet file SecretSeed.md that explains what this is and why to use it, and also how to reset the login password you have not included it, with instructions to check all system settings and test to make sure everything works after uploading the config file without it.
It updates content in several articles that include the secret seed information.
It fixes spelling issues and updates the GeneralSettingsScreens.md UI ref article. It changes non-UI ref content from tutorial style to UI ref style.

* PD-1683 Style fixes and Field Correction

This commit fixes a mis-named field, and makes a style change to the UpdateHASCALE.md article to indent the snippet to align it with the numbered list.

* Update content/SCALE/SCALEUIReference/SystemSettings/GeneralSettingsScreens.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update static/includes/SecretSeed.md

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>

* Update UpdateScreensSCALE.md

Reword lines 64 and 65 for clarity

* Update ManageSysConfigSCALE.md

Reworded the warning in lines 34-40 and added a second warning hint admonition about the secret seed in lines 42-46.

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>
Co-authored-by: linzibelle <linzi.moore@ixsystems.com>