PD-1860 / 25.10 / PD-1860 Update Reporting UI Ref (#3680)
* PD-1860 Update Reporting UI Ref

This PR updates the content in the Reporting UI ref article, removes the incorrect information, and makes a few style guide corrections.

* PD-1860 Remove Netdata references

This commit removes the references to Netdata that was removed from TrueNAS a few releases ago. It deletes the netdatareporting.md snippet file.

* Revert "PD-1860 Remove Netdata references"

This reverts commit 1ca82e42183be4ddb6fbc8eb306e6f01e3e451f7.

* PD-1860 Revert Commit Removing NetdataReporting.md Snippet

Reverting commit removing the netdatareporting.md snippet.

* PD-1860 Update Screenshots

This commit updates screenshots with 25.04 versions

* PD-1860 Update Reporting UI Ref

This commit removes outdated reporting images that no longer exist in the UI

* Update ReportingScreensSCALE.md

* PD-1860 Reword ZFS ARC Image Caption

This commit rewords the name of the image in the ZFS ARC section, and deletes the second image.

---------

Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>