PD-1862 / 25.04 / Pd 1862 25.04 (#3707)
* Empty commit to create PR on github.

You should reset it

* PD-1862 / 25.10 / Pd 1862 create enterprise ad content (#3678)

* PD-1862 Add new and modify existing snippet

This commit adds a new snippet AddPrivilege.md with the procedure to add and define a new privilege, and adds it to the existing AccessSettingsWidget.md snippet.
It adds new screenshots, and updates the LocalGroupsScreens.md UI ref article with style guide corrections

* PD-1862 Create another new snippet

This commit moves the procedure content into a new ConfigureAccessSettings.md snippet and replaces the duplicated content in the AdvancedSettingsScreen.md UI ref article with the AccessSettingsWidget.md snippet that covers what is shown in the widget but not the procedure information now in the ConfigureAccessSettings.md snippet.

* PD-1862 Rename AccessSettingsWidget

This commit renames the AccessSettingsWidget.md to AccessWidget.md

* PD-1862 Text and instruction changes

This commit makes additional changes to tutorial and snippet content.
Updated the ConfigADSCALE.md instructions to include network changes to make as part of the process.
Changes to the AddPrivilege.md snippet, correcting a few steps.
Runs spell and grammar checks.

* Update content/SCALE/SCALETutorials/Credentials/DirectoryServices/ConfigADSCALE.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/SystemSettings/AdvancedSettingsScreen.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update static/includes/AddPrivilege.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update static/includes/AddPrivilege.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>

* Update ManageLocalGroups.md

Reword bullet list in lines 87-89

* Update ConfigureAccessSettings.md

Reword bullet list in lines 36-38

---------

Co-authored-by: linzibelle <linzi.moore@ixsystems.com>
(cherry picked from commit 80d4d11f4cb831a70827ecae0d8d7adf4b320ae1)

* PD-1862 Fix Typos and Enterprise Shortcode Postionion

This PR fixes the position of the enterprise shortcode to encircle line 26 and rewords line 26.

It also fixes typos for the word execute.

* Update ManageLocalGroups.md

---------

Co-authored-by: Bugclerk <bugclerk@ixsystems.com>
Co-authored-by: DjP-iX <133042991+DjP-iX@users.noreply.github.com>