PD-1924 / 25.10 / Pd 1924 update group privileges screen content (#3728)
* PD-1924 Update Groups UI Ref Article

This PR updates the Add Privilege sreenshot and updates the setting descriptions to reflect the change, removing the Directory Services Groups field.

* PD-1924 Update the Group Screen UI ref Article Privileges Content

This PR updates the Add Privilege screenshot and setting descriptions, and links it to the Access widget setting for Allowing Directory Service users to access WebUI option as this influences what Enterprise customers see and experience.

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>

* Update content/SCALE/SCALEUIReference/Credentials/LocalGroupsScreens.md

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>

* Update LocalGroupsScreens.md

Verify commit, add carriage return in line 84 at the second sentence in the paragraph

---------

Co-authored-by: Tim Moore <Mrt134@users.noreply.github.com>