PD-2046: update Getting Support article in 24.10 (#3911)
There's some elements that aren't being updated or are specific to 24.10. Backports to other branches will need to be reviewed to ensure the content changes are appropriate for their TrueNAS version.

Reorganize content a bit, make sure links are functional, and update forum snippet to note it's use for feature suggestions and voting.

Co-authored-by: MicJ <92740932+micjohnson777@users.noreply.github.com>